# Optimization pass - parallelize parsing and IR compilation, less expensive string ops