--- a/src/main/java/dev/m0rg/howl/ast/ImportStatement.java
+++ b/src/main/java/dev/m0rg/howl/ast/ImportStatement.java
@@ -25,7 +25,7 @@ public class ImportStatement extends ASTElement {
         ;
     }
 
-    public String getPath() {
+    public String getImportPath() {
         return path;
     }
 }