--- a/src/main/java/dev/m0rg/howl/ast/NameHolder.java
+++ b/src/main/java/dev/m0rg/howl/ast/NameHolder.java
@@ -6,14 +6,13 @@ import java.util.Optional;
 public interface NameHolder {
     Optional<ASTElement> getChild(String name);
 
-    default Optional<ASTElement> getPath(String path) {
-        String[] parts = path.split("\\.");
+    default Optional<ASTElement> getPath(String[] parts) {
         if (parts.length == 1) {
             return this.getChild(parts[0]);
         } else {
             return this.getChild(parts[0]).flatMap(x -> {
                 if (x instanceof NameHolder) {
-                    return ((NameHolder) x).getPath(String.join(".", Arrays.copyOfRange(parts, 1, parts.length)));
+                    return ((NameHolder) x).getPath(Arrays.copyOfRange(parts, 1, parts.length));
                 } else {
                     return Optional.empty();
                 }