```
morgan@Morgans-MBP ~/src/howl-lang (main?) $ time java -jar target/howlc-1.0-SNAPSHOT.jar unit_tests/iterator.hl -o /tmp/test
java -jar target/howlc-1.0-SNAPSHOT.jar unit_tests/iterator.hl -o /tmp/test  6.19s user 0.50s system 261% cpu 2.560 total
morgan@Morgans-MBP ~/src/howl-lang (main?) $ git switch transform-optimization
Switched to branch 'transform-optimization'
morgan@Morgans-MBP ~/src/howl-lang (transform-optimization) $ make >/dev/null 2>&1
morgan@Morgans-MBP ~/src/howl-lang (transform-optimization) $ time java -jar target/howlc-1.0-SNAPSHOT.jar unit_tests/iterator.hl -o /tmp/test
java -jar target/howlc-1.0-SNAPSHOT.jar unit_tests/iterator.hl -o /tmp/test  3.74s user 0.44s system 377% cpu 1.107 total
```